refactoring DynamicTabAdapter extract ConfirmCounter logic
