+++ app/src/main/java/org/eyeseetea/malariacare/layout/adapters/survey/DynamicTabAdapter.java
@@ -309,77 +309,6 @@ public class DynamicTabAdapter extends BaseAdapter implements ITabAdapter {
         isClicked = false;
     }
 
-    public void showStandardConfirmCounter(final View view, final Option selectedOption,
-            final Question question,
-            Question questionCounter) {
-        //Change question x confirm message
-        View rootView = view.getRootView();
-        final TextCard questionView = (TextCard) rootView.findViewById(R.id.question);
-        questionView.setText(questionCounter.getInternationalizedForm_name());
-        ProgressUtils.setProgressBarText(rootView, "");
-        //cancel
-        ImageView noView = (ImageView) rootView.findViewById(R.id.confirm_no);
-        noView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //Leave current question as it was
-                removeConfirmCounter(v);
-                notifyDataSetChanged();
-                isClicked = false;
-            }
-        });
-
-        //confirm
-        ImageView yesView = (ImageView) rootView.findViewById(R.id.confirm_yes);
-        yesView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                navigationController.increaseCounterRepetitions(selectedOption);
-                removeConfirmCounter(v);
-                saveOptionValue(view, selectedOption, question, true);
-            }
-        });
-
-        //Show confirm on full screen
-        rootView.findViewById(R.id.no_scrolled_table).setVisibility(View.GONE);
-        rootView.findViewById(R.id.scrolled_table).setVisibility(View.GONE);
-        rootView.findViewById(R.id.confirm_table).setVisibility(View.VISIBLE);
-
-        //Show question image in counter alert
-        if (questionCounter.getPath() != null && !questionCounter.getPath().equals("")) {
-            ImageView imageView = (ImageView) rootView.findViewById(R.id.questionImageRow);
-            BaseLayoutUtils.putImageInImageView(questionCounter.getInternationalizedPath(),
-                    imageView);
-            imageView.setVisibility(View.VISIBLE);
-        }
-
-        //Question "header" is in the first option in Options.csv
-        List<Option> questionOptions = questionCounter.getAnswer().getOptions();
-        if (questionOptions.get(0) != null) {
-            TextCard textCard = (TextCard) rootView.findViewById(R.id.questionTextRow);
-            textCard.setText(questionOptions.get(0).getInternationalizedCode());
-            textCard.setTextSize(questionOptions.get(0).getOptionAttribute().getText_size());
-        }
-        //Question "confirm button" is in the second option in Options.csv
-        if (questionOptions.get(1) != null) {
-            TextCard confirmTextCard = (TextCard) rootView.findViewById(R.id.textcard_confirm_yes);
-            confirmTextCard.setText(questionOptions.get(1).getInternationalizedCode());
-            confirmTextCard.setTextSize(questionOptions.get(1).getOptionAttribute().getText_size());
-        }
-        //Question "no confirm button" is in the third option in Options.csv
-        if (questionOptions.get(2) != null) {
-            TextCard noConfirmTextCard = (TextCard) rootView.findViewById(R.id.textcard_confirm_no);
-            noConfirmTextCard.setText(questionOptions.get(2).getInternationalizedCode());
-            noConfirmTextCard.setTextSize(questionOptions.get(
-                    2).getOptionAttribute().getText_size());
-        }
-    }
-
-    public void removeConfirmCounter(View view) {
-        view.getRootView().findViewById(R.id.dynamic_tab_options_table).setVisibility(View.VISIBLE);
-        view.getRootView().findViewById(R.id.confirm_table).setVisibility(View.GONE);
-    }
-
     private void darkenNonSelected(View view, Option selectedOption) {
         swipeTouchListener.clearClickableViews();
         //A Warning or Reminder (not a real option)
@@ -807,14 +736,6 @@ public class DynamicTabAdapter extends BaseAdapter implements ITabAdapter {
         Float maxCounter = selectedOption.getFactor();
 
         return counterValue.equals(maxCounter);
-    }
-
-    /**
-     * Used to set the text widht like the framelayout size
-     * to prevent a resize of the frameLayout if the textoption is more bigger.
-     */
-    private void resizeTextWidth(FrameLayout frameLayout, TextCard textOption) {
-        textOption.setWidth(frameLayout.getWidth());
     }
 
     /**
